+++ packages/account-eip6963/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @sophon-labs/account-eip6963
 
+## 2.1.5
+
+### Patch Changes
+
+- Changing initialization order on rn
+- Updated dependencies
+  - @sophon-labs/account-provider@2.1.5
+  - @sophon-labs/account-core@2.1.5
+
+## 2.1.4
+
+### Patch Changes
+
+- Cleanup on authentication on fresh install
+- Updated dependencies
+  - @sophon-labs/account-provider@2.1.4
+  - @sophon-labs/account-core@2.1.4
+
+## 2.1.3
+
+### Patch Changes
+
+- Removing disconnect from react native and adding is initialized
+- Updated dependencies
+  - @sophon-labs/account-provider@2.1.3
+  - @sophon-labs/account-core@2.1.3
+
 ## 2.1.2
 
 ### Patch Changes